requirements.txt
awscli
aws_cdk.core
aws-cdk.aws-apigateway
aws-cdk.aws-cognito
aws-cdk.aws-s3-notifications
aws-cdk.aws-sqs
aws_cdk.aws_ec2
aws_cdk.aws_iam
aws_cdk.aws_elasticsearch
aws_cdk.aws_lambda
aws_cdk.aws_s3
boto3
